Movie detection Zombie to SchedulerJob
